# PERF: halve render time in Styler with itertuples