@jreback 2021-02-22
great. Can you add a whatsnew note (IIRC you may already have one about perf rendering, if so just add this on). ping on greenish.