@attack68 2021-02-23
@jreback ping greenish. `test_from_coo` failure is unrelated.